+++ src/compas_timber/fabrication/joint_factories/t_butt_factory.py
@@ -1,6 +1,5 @@
 from compas_timber.connections import TButtJoint
 from compas_timber.fabrication import BTLx
-from compas_timber.fabrication.btlx import BTLxPart
 from compas_timber.fabrication import BTLxJackCut
 from compas_timber.fabrication.btlx_processes.btlx_lap import BTLxLap
 from compas_timber.fabrication.btlx_processes.btlx_double_cut import BTLxDoubleCut